\name{plotHarmonics}
\alias{plotHarmonics}

\title{Create Continuous Assumption Plots}

\description{
plotHarmonics creates plots that help assess the continuousness assumption of interactions within the
standard multiplicative model. Each plot assesses interaction beta coefficient magnitude and
statistical significance across a number of percentile-ranked regions of the moderator.
}

\usage{
plotHarmonics(X, Y, M, data = NULL, nbin = 3, smooth = FALSE, minmax = NULL, labels = NULL, sim = NULL, alpha = .05, main = NULL)
}

\arguments{
  \item{X}{The predictor in the linear interaction model.}
  \item{Y}{The outcome in the linear interaction model.}
  \item{M}{The moderator (or second predictor) in the linear interaction model.}
  \item{data}{Dataset containing X, Y, and M. The default is NULL and requires another value if a simulation dataset is not defined.}
  \item{nbin}{The maximum number of percentile-ranked regions of M to create within sets. The default is 3 regions.}
  \item{smooth}{If TRUE, lines are a smoothed spline in Plot 3. If FALSE, lines are step-wise between regions. The default = FALSE.}
  \item{minmax}{Defines the Y-axis scale for Plot 3 with a format of c(minimum value, maximum value). The default is NULL. If NULL, the scale is automatically defined.}
  \item{labels}{Defines the X- and Y-axis titles for Plot 3 with a format of c("X-axis title", "Y-axis title"). The default is NULL. If NULL, the titles are automatically defined.}
  \item{sim}{Simulated dataset containing X, Y, and M, as generated by the bootCA function. The default is NULL and requires another value if a real-world dataset is not defined.}
  \item{alpha}{The alpha criteria for two-sided statistical tests. The default is .05.}
}
  \item{main}{If a simulated dataset is designated, assigns title of figure 2. The default is NULL.}
}

\value{
  \item{Plot 1}{Plot 1 describes the percentage of regions in which the interaction coefficient is statistically
significant across sets of regions.}
  \item{Plot 2}{Plot 2 describes the p-value of interactions in moderator regions across sets. This plot is only generated for real-world datasets.}
  \item{Plot 3}{Plot 3 describes the interaction magnitudes in moderator regions over values of the moderator.}
}

\author{
Shane J. Sacco, M.A.
University of Connecticut
}

\section{Warning }{Please note for nbin: values less than the default of 3 regions are required for nbin if the sample size of regions reaches n <= 25.}

\examples{
  Predictor <- rnorm(1500, 0, 0.5)
  Moderator <- rnorm(1500, 0, 0.5)
  Outcome <- rep(0, 1500)
  Error <- rnorm(1500, 0, 0.5)
  Outcome [Moderator <= 0] = 0.5 * Predictor [Moderator <= 0] + Error [Moderator <= 0]
  Outcome [Moderator > 0]  = -0.5 * Predictor [Moderator > 0] + Error [Moderator > 0]

  ExampleDat <- data.frame(Predictor, Moderator, Outcome)

  plotHarmonics (X = ExampleDat$Predictor, Y = ExampleDat$Outcome, M = ExampleDat$Moderator, data=ExampleDat, smooth = FALSE, nbin = 3, minmax = c(-3,3), labels = c("Moderator (percentile)", "Interaction Beta"), alpha = .05)
}

\keyword{ hplot }
\keyword{ distribution }
